requirements.txt
discord.py==1.7.3
matplotlib==3.4.3
numpy==1.21.2
sqlite3==2.6.0
